requirements.txt
python-pptx==1.0.2
tqdm==4.66.1
pydantic==2.6.1
sentence-transformers==2.7.0
faiss-cpu==1.9.0.post1
